Move `components` -> `elements`
