+++ packages/avatars/src/elements/Avatar.spec.tsx
