+++ packages/avatars/src/index.ts
@@ -5,4 +5,4 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-export { default as Avatar } from './components/Avatar';
+export { default as Avatar } from './elements/Avatar';
